requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.42
botocore==1.42.42
certifi==2026.1.4
click==8.3.1
dnspython==2.8.0
fastapi==0.128.0
h11==0.16.0
idna==3.11
jmespath==1.1.0
motor==3.7.1
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
s3transfer==0.16.0
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
